+++ src/lib/Client/index.ts
@@ -19,3 +19,4 @@ export * from './Common';
 export * from './TCPClient';
 export * from './HttpClient';
 export * from './HttpsClient';
+export * from './BuiltInRIDGenerator';